+++ .changeset/sixty-clouds-confess.md
@@ -2,4 +2,4 @@
 '@shopify/theme-check-common': minor
 ---
 
-New check: ValidVisibleIf
+Add `ValidVisibleIf` check
